chore: clean up, refactor, fix eslint
