+++ src/server/botd-firewall/blockedIpsDatabase.ts
@@ -1,6 +1,6 @@
 import { Attributes, DataTypes, InferAttributes, InferCreationAttributes, Model } from 'sequelize';
 import { sequelize } from '../server';
-import { MAX_BLOCKED_IPS } from './updateFirewallRule';
+import { MAX_BLOCKED_IPS } from './buildFirewallRules';
 
 interface BlockedIpAttributes
   extends Model<InferAttributes<BlockedIpAttributes>, InferCreationAttributes<BlockedIpAttributes>> {
